Move helper functions over to test_helpers.go
